add modale setting mobile
