+++ src/components/ModalWindow/ModalWindow.module.css
@@ -2,7 +2,6 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  position: fixed;
   top: 0;
   left: 0;
   bottom: 0;
@@ -33,14 +32,11 @@
   background-color: transparent;
   position: absolute;
   z-index: 999999;
-  top: 10px;
-  right: 10px;
+  top: 16px;
+  right: 12px;
   cursor: pointer;
 }
 
 .closeIcon {
-  padding: 0;
-  width: 32px;
-  height: 32px;
   fill: var(--main-black-color);
 }
