+++ src/components/UserPanel/UserPanel.module.css
@@ -67,6 +67,8 @@
 .avatarImg {
   border-radius: 50%;
   object-fit: cover;
+  width: 100%;
+  height: 100%;
 }
 
 @media screen and (min-width: 768px) {
